+++ vbr/vbrCore/functions/Q_method_units.m
@@ -1,0 +1,9 @@
+function units = Q_method_units()
+    units.J1 = '';
+    units.J2 = '';
+    units.Q = '';
+    units.Qinv = '';
+    units.M = 'Pa'; 
+    units.V = 'm/s';
+    units.Vave = 'm/s';
+end