+++ vbr/vbrCore/functions/sr_visc_calc_HZK2011.m
@@ -56,5 +56,9 @@ function VBR= sr_visc_calc_HZK2011(VBR)
   % store total composite strain rate and effective viscosity
   VBR.out.viscous.HZK2011.sr_tot=sr_tot; % total strain rate
   VBR.out.viscous.HZK2011.eta_total = sig*1e6./sr_tot ; % total viscosity
-
+  units.sr = "1/s";
+  units.eta = "Pa*s";
+  units.sr_tot = "Pa*s";
+  units.eta_tot = "Pa*s";
+  VBR.out.viscous.HZK2011.units = units;
 end