+++ release_notes.md
@@ -1,10 +1,10 @@
-# v0.99.5 
+# v1.0.0
 
-Release summary for v0.99.5.
+This is the first series-1 release! It is backwards compatible.
 
 ## new features 
 
-(add any new features here)
+- units metadata: all output methods now contain a units structure that contains the units for each output field. You can also call `SV_input_units()` to get a structure that lists the expected units for each `VBR.in.SV` field.
 
 ## bug fixes